removed UserDetailsForm and added logic on Same as User to button to diplay only for log in users
